==> Code/Step1_DataRetrieval_Mongo.ipynb <==
#Connect to MongoDB
db.createUser({
    user: 'UsernameHidden',
    pwd: 'PasswordHidden',
    roles: [{ role: 'readWrite', db:'project4'}]
})

import pymongo

client = pymongo.MongoClient("mongodb://username:password@IPAddress/project4") # defaults to port 27017

db = client.project4

db = client.project4
data = db.royaltweets

df = pd.DataFrame(data)

#pickel the dataframe
df.to_pickle('mongodatabase.pkl')
