==> Code/Step2_DataCleaning_TopicModeling_Vectorization.ipynb <==
import pandas as pd
import re
import html.parser
from sklearn.feature_extraction.text import CountVectorizer
import pickle
from nltk.corpus import stopwords
from nltk.stem import porter
import nltk
import pickle
from sklearn.decomposition import NMF, TruncatedSVD, LatentDirichletAllocation

stemmer = porter.PorterStemmer()

stopwords = stopwords.words()

f = open('mongodatabase.pkl','rb')
df = pickle.load(f)

added_stopwords = ['wait','cant','bi','hi','always','long','short','getting','away','see','watching','one','better','today','get','th',
'could','feel','would','hours','must','watch','st','theres','much','probably','came','brought','lot','last','bit',
'youre','need','make','piece','ass','nobody','thought','cosa','think','right','behind','late','bringing','facupfinal',
'cup','troian','half','find','really','many','never','oh','actual','ser','ways','stream','via','name','street','vote',
'something','nothing','still','top','set','didnt','us','along','two','lots','everyone','shes','turn','well','high',
'anyone','says','become','ten','mo','gets','click','gave','give','knew','forward','ver','voy','definitely','haha',
'sat','join','old','fucking','hrh','ms','sure','gonna','whatever','making','since','price', 'got','may','let','litte',
'aqu','without','year old','big','sometimes','nd','look','got']

stopwords = stopwords + added_stopwords

df.tail(5)

def one_tweet_clean(tweet):
    punctuations = '''!()-[]{};:'"\,<>./?$%^&*_~'''
    new_doc = re.sub(r'http\S+', '', tweet)
    new_doc = html.unescape(new_doc)
    new_doc = new_doc.replace("\n", "")
    new_doc = new_doc.replace("\\","")
    new_doc = new_doc.replace("@Suits_USA","Suits_USA")
    new_doc = new_doc.replace("@RoyalFamily","RoyalFamily")
    new_doc = new_doc.replace("@BishopCurry","BishopCurry")
    new_doc = new_doc.replace("@KensingtonRoyal","KensingtonRoyal")
    new_doc = new_doc.lower()
    new_doc = re.sub(r'@\w+','',new_doc)
    new_doc = re.sub(r'[0-9]', '',new_doc)
    new_doc = stemmer.stem(new_doc)
    for punc in punctuations:
        new_doc = new_doc.replace(punc, "")
    if new_doc in stopwords:
        return None
    return new_doc

df['cleaned_text'] = df['text'].apply(lambda x: one_tweet_clean(x))

df.to_pickle('updateddataframe.pkl')

df.head(5)

def clean_first(text):
    new_text = []
    punctuations = '''!()-[]{};:'"\,<>./?$%^&*_~'''
    for doc in text:  
        new_doc = one_tweet_clean(doc)
        if doc is not None:   
            new_text.append(new_doc)
    return new_text

docs = df['text'].values.tolist()

new_docs = clean_first(docs)

with open('cleandata.pkl','wb') as file:
    pickle.dump(new_docs, file)

with open('cleandata.pkl','rb') as file:
    new_docs = pickle.load(file)

# takes out Chinese, Japanese, and Korean characters with unicode data
def emoticon_tokenize(s):
    token_pattern = r'(?<=^|(?<=\b))[\w\#]\w+\b'
    emoji_regex = re.compile(r'[\u263a-\U0001f645]')
    token_regex = re.compile(token_pattern, flags=re.ASCII)
    emoticons = emoji_regex.findall(s)
    parsed_emoticons=[] 
    for e in emoticons: 
        try: 
            if unicodedata.name(e)[:3]!='CJK' and unicodedata.name(e)[:8]!='HIRAGANA' and unicodedata.name(e)[:8]!='KATAKANA':
                parsed_emoticons.append(e)
        except: 
            pass
    word_list = token_regex.findall(s)
    return word_list + parsed_emoticons

%%time
cv = CountVectorizer(ngram_range=(1, 3), tokenizer = emoticon_tokenize, max_features=3000, stop_words=stopwords, max_df = 0.6)
cv.fit(new_docs)
len(cv.vocabulary_)

with open('cv.pkl','wb') as file:
    pickle.dump(cv, file)

cv.vocabulary_

cv_data = cv.transform(new_docs)

def display_topics(model, feature_names, num_top_words, topic_names=None):  
    for ix, topic in enumerate(model.components_):  #model.components_ are topic probability distributions
        if not topic_names or not topic_names[ix]:
            print("\nTopic ", ix)
        else:
            print("\nTopic: '",topic_names[ix],"'")
        print(", ".join([feature_names[i]
                        for i in topic.argsort()[:-num_top_words - 1:-1]]))

n_comp = 5   #number of topics
lsa_cv = TruncatedSVD(n_components=n_comp)
nmf_cv = NMF(n_components=n_comp)

# Try the LDA model as well
lda_cv = LatentDirichletAllocation(n_components=n_comp)

%%time
#trained models. so lsa_cv is already fitted and transformed. Think of lr.fit() --> lr.coef_
# going from big document space to feature space. Think about rabbit and dish

lsa_cv_data = lsa_cv.fit_transform(cv_data)

nmf_cv_data = nmf_cv.fit_transform(cv_data)

with open('nmf_cv.pkl','wb') as file:
    pickle.dump(nmf_cv, file)

with open('lsa_cv.pkl','wb') as file:
    pickle.dump(lsa_cv, file)

%%time
lda_cv_data = lda_cv.fit_transform(cv_data)

# display top 5 features of the lsa model
display_topics(lsa_cv,cv.get_feature_names(),30)   #10 is the number of top words

# display top 5 features of the nmf model
display_topics(nmf_cv,cv.get_feature_names(),30)

# display top 5 topics for the lda model
display_topics(lda_cv, cv.get_feature_names(),30)
