==> Code/Step3_LocationProcessing.ipynb <==
import pickle
import pandas as pd
import numpy as np

f = open('updateddataframe.pkl','rb')
df = pickle.load(f)

df.head(5)

df['loc'] = df['location'].str.lower()

possible_locations = {'uk':['manchester','london','glasgow','leeds','birmingham','yorkshire','cardiff', 'bristol',
                           'nottingham','sheffield','northern ireland','belfast, northern ireland', 'kent',
                           'brighton', 'belfast', 'newcastle','kensington palace','leicester','cheshire','221b baker street',
                           'oxford','whiltshire','cambridge','Yorkshire and The Humber','windsor'],
                      'usa':['san diego','alexandria, va','arlington, va','colorado','indiana','arizona','dallas, texas',
                             'durham, nc','long island, ny','nc','scottsdale, az','jersey','houston','fort lauderdale, fl',
                             'jacksonville, fl','charleston, sc','dc','beverly hills, ca','baton rouge, la', 'u.s.a','alabama',
                             'west hollywood, ca','greenville, sc','illinois','tuscan, az','santa monica, ca','socal',
                             'ca','honolulu, hi','america','long beach, ca','oregon','omaha, ne','los angeles','pennsylvania',
                             'bay area','minneapolis','atl','pa','kentucky','dallas','st. louis, mo','nyc','grand rapids, mi',
                             'modesto, california','hiuntsville, al','fl','southampton','columbia, sc','virginia beach, va',
                             'wisconsin','orange county, ca','tennessee','tx',
                          'los angeles, ca','new york, ny', 'new york city','washington, dc','atlanta, ga', 
                             'chicago, il', 'new york','boston, ma', 'los angeles','nyc', 'houston, tx', 
                             'brooklyn, ny','philadelphia, pa', 'dallas, tx', 'chicago', 'miami, fl', 'san francisco, ca',
                            'seattle, wa','austin, tx','california','nashville, tn','texas','orlando, fl','charlotte,nc',
                            'las vegas, nv', 'san diego, ca','florida','manhattan, ny','denver, co','detroit, mi',
                             'portland, or','pittsburg, pa','columbus, oh','tampa, fl','washington dc','san antonio, tx',
                            'new orleans, la','minneapolis, mn','washington, d.c.','puerto rico','cleveland, oh',
                            'raleigh, nc','north carolina','hollywood, ca','boston','cincinnati, oh','ohio','san francisco',
                            'brooklyn','louisville, ky','memphis, tn','milwaukee, wi','michigan','southern california',
                            'new york, new york', 'virginia','seattle','nj','maryland','kansas city, mo','georgia',
                            'los angeles, california','buffalo, ny','sacramento, ca','ny','richmond, va','miami',
                            'st louis, mo','la','knoxville, tn','birmingham, al','rochester, ny','oakland, ca',
                            'cleveland, ohio','galt, ca', 'massachusetts','minnesota','south florida','fort worth, tx',
                            'houston, tx','bronx, ny', 'san jose, ca','san francisco bay area','queens, ny','south caorlina',
                            'salt lake city, ut','chattanooga, tn','west palm beach, fl','connecticut','austin, texas',
                            'las vegas']}

possible_countries = possible_locations.keys()

def assign_countries(location):
    if not location: 
        return np.nan
    if 'usa' in location:
        return 'usa'
    if 'us' in location:
        return 'usa'
    if 'united states' in location:
        return 'usa'
    if 'england' in location:
        return 'uk'
    if 'united kingdom' in location:
        return 'uk'
    if 'uk' in location:
        return 'uk'
    if 'scotland' in location:
        return 'uk'
    if 'wales' in location:
        return 'uk'
    if 'london' in location:
        return 'uk'

    for country in possible_countries:
        if location in possible_locations[country]:
            return country
    if not None: 
        return 'other'

df['country'] = df['loc'].apply(assign_countries)

df.head(5)

df.to_pickle('locationdataframe.pkl')
